color_grid: initial copy-paste implementation of image write, no need for stb_image and simple fix in file reading.

+++ color_grid/src/include/read.h
@@ -2,9 +2,10 @@
 #define _READ_H_INCLUDED_
 
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 
 // return true on success and false on failure.
-bool read_file_for_hex_colors(const char *path, uint32_t *out, uint32_t out_length);
+bool read_file_for_hex_colors(const char *path, uint32_t *out, size_t out_length);
 
 #endif
